# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Fixit 24/7 API - On-demand home repair marketplace
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List jobs (role-filtered)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: urgency
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Job list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobListResponse"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Post a new job (homeowner)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobBody"
      responses:
        "201":
          description: Job created and matching triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get job details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateJob
      tags: [jobs]
      summary: Update a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateJobBody"
      responses:
        "200":
          description: Updated job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Delete/cancel a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # Job Claims
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: Get user notifications
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Notifications list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    put:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [homeowner, tradie, admin]
        phone:
          type: string
          nullable: true
        suburb:
          type: string
          nullable: true
        postcode:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        rating:
          type: number
          nullable: true
        reviewCount:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - reviewCount
        - isActive
        - createdAt
...
    Job:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [open, matched, in_progress, completed, cancelled]
        urgency:
          type: string
          enum: [standard, urgent, emergency]
        categoryId:
          type: integer
        categoryName:
          type: string
          nullable: true
        homeownerId:
          type: integer
        homeownerName:
          type: string
          nullable: true
        suburb:
          type: string
          nullable: true
        postcode:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        imageUrls:
          type: array
          items:
            type: string
        budget:
          type: number
          nullable: true
        claimCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        scheduledFor:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - title
        - description
        - status
        - urgency
        - categoryId
        - homeownerId
        - imageUrls
        - claimCount
        - createdAt
    JobDetail:
      allOf:
        - $ref: "#/components/schemas/Job"
        - type: object
          properties:
            claims:
              type: array
              items:
                $ref: "#/components/schemas/Claim"
            matchedTradies:
              type: array
              items:
                $ref: "#/components/schemas/User"
    JobListResponse:
      type: object
      properties:
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/Job"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - jobs
        - total
        - page
        - limit
    CreateJobBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        categoryId:
          type: integer
        urgency:
          type: string
          enum: [standard, urgent, emergency]
        suburb:
          type: string
        postcode:
          type: string
        address:
          type: string
        imageUrls:
          type: array
          items:
            type: string
        budget:
          type: number
        scheduledFor:
          type: string
          format: date-time
      required:
        - title
        - description
        - categoryId
        - urgency
    UpdateJobBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [open, matched, in_progress, completed, cancelled]
        urgency:
          type: string
          enum: [standard, urgent, emergency]
        address:
          type: string
        budget:
          type: number
        scheduledFor:
          type: string
          format: date-time
    Claim:
      type: object
      properties:
        id:
          type: integer
        jobId:
          type: integer
        tradieId:
          type: integer
        tradieName:
          type: string
          nullable: true
        tradieRating:
          type: number
          nullable: true
        tradieReviewCount:
          type: integer
          nullable: true
        tradieSuburb:
          type: string
          nullable: true
        tradieAvatarUrl:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, accepted, rejected, withdrawn, completed]
        message:
          type: string
          nullable: true
        proposedPrice:
          type: number
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - jobId
        - tradieId
        - status
        - createdAt
        - tradieReviewCount
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
        title:
          type: string
        message:
          type: string
        isRead:
          type: boolean
        jobId:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - type
        - title
        - message
        - isRead
        - createdAt
...